requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
cachetools==4.2.4
celery==5.3.4
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dj-config-url==0.1.1
dj-database-url==2.1.0
Django==4.2.5
django-ckeditor==6.7.1
django-cors-headers==4.2.0
django-debug-toolbar==4.3.0
django-filter==23.2
django-jazzmin==3.0.1
django-js-asset==2.2.0
django-jsonform==2.22.0
django-modeltranslation==0.18.11
django-querycount==0.8.3
django-ratelimit==4.1.0
django-redis==5.4.0
django-restful-translator==0.1.4
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
Faker==24.4.0
google-api-core==1.34.1
google-auth==1.35.0
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
googleapis-common-protos==1.63.0
googletrans==4.0.0rc1
grpcio==1.62.1
grpcio-status==1.48.2
gunicorn==21.2.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.4.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.7
packaging==24.0
Pillow==10.0.0
polib==1.2.0
prompt_toolkit==3.0.46
protobuf==3.20.3
psycopg2==2.9.10
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
referencing==0.34.0
requests==2.31.0
rfc3986==1.5.0
rpds-py==0.18.0
rsa==4.9
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
toposort==1.10
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
django-dirtyfields==1.9.5
psycopg2-binary==2.9.10